core, internal: move workerpool into syncx
